fix : type of seeds

+++ src/main/java/com/example/kakao/dto/Image.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class Image {
     private String id;
     private String image;
-    private int seed;
+    private long seed;
     private Object nsfw_content_detected; // null을 처리하기 위해 Object를 사용
     private Object nsfw_score; // null을 처리하기 위해 Object를 사용
 }
